Update appender tests to reflect computed columns support
Prior to https://github.com/duckdb/duckdb/pull/14346, the C API did not
support appending to tables with computed columns.

This would cause panics until
https://github.com/duckdb/duckdb-rs/pull/296, where a test was
introduced asserting that an append to a table with computed columns
returns an error value.

Now that appending to is supported, update the test to reflect that.