[WIP] Upgrade to V1.2.1 (#462)
* fix: pin duckdb and libduckdb-sys versions

Signed-off-by: luong-hq <luong@hyperquery.ai>

* Bump arrow to the version compatible with chrono 0.4.40

* Update only arrow

* Fix minor typo

* Do not pass stdlib

* Fix #380, bug in code that reads Enums

In turning a ValueRef to a Value, the code was indexing with the row index directly into the dictionary's string array, so iterating any query's results was effectively just listing the Enum variants in order until row index > number of variants and it panicked. Fix by getting the row value (key) and using that to index into the dictionary's string array.

* Fix incorrect Enum test

The test tests for correct value extraction, based on built-in examples in DuckDB. The old expected value was flatly incorrect, as can be seen in duckdb directly with `SELECT medium_enum FROM test_all_types();`.

* Fix writing Arrow Decimal128 with zero scale

* Fix CI install of rustfmt and clippy

Extra space breaks hecrj/setup-rust-action. Arguments should be comma-separated.

* update duckdb to v1.2.1

* replace rust toolchain action

* fix warnings

* add back sanitizer

* move logicaltypeid

* add correct component for asan

* add back extensions to asan build

* bundle ICU as part of extensions-full

---------

Signed-off-by: luong-hq <luong@hyperquery.ai>
Co-authored-by: luong-hq <luong@hyperquery.ai>
Co-authored-by: Hiroaki Yutani <yutani.ini@gmail.com>
Co-authored-by: Shritesh Bhattarai <shr@ite.sh>
Co-authored-by: Nickolas Fotopoulos <nickolas.fotopoulos@gmail.com>
Co-authored-by: sgrebnov <sergei.grebnov@gmail.com>