Implement conversion to ValueRef for list values
`Value::List` is backed by a `Vec<Value>`, however, `ValueRef::List`
can't refer to it because `ValueRef` assumes referee is backed by Arrow
list arrays.

This commit extends `ValueRef::List` to allow refering to native Rust
values that are not backed by Arrow list arrays.

Note that this is a user-facing braking change in `ValueRef::List` and
`ListType` API.